requirements.txt
Flask==1.1.2
gunicorn==20.0.3
numpy==1.19.2
opencv-python-headless==4.2.0.32
#opencv-python==4.0.0.21 #python version issue :FIXME
Pillow==7.1.2
python-dotenv==0.13.0
scikit-learn==0.19.1
scipy==1.5.2
